// src/models/deviceType.ts
export type DeviceType =
  | "UNKNOWN"
  | "BAG_TAG_PRINTER"
  | "BAG_TAG_FEEDER"
  | "BAG_TAG_DISPENSER"
  | "BOARDING_PASS_PRINTER"
  | "BOARDING_PASS_FEEDER"
  | "BOARDING_PASS_DISPENSER"
  | "BARCODE_READER"
  | "PASSPORT_READER"
  | "MSR_READER"
  | "KEY_PAD"
  | "ANNOUNCEMENT"
  | "FEEDER"
  | "DISPENSER"
  | "SCALE"
  | "BIOMETRIC"
  | "CAMERA"
  | "ILLUMINATION"
  | "HEADSET"
  | "INSERTION_BELT"
  | "VERIFICATION_BELT"
  | "PARKING_BELT"
  | "RFID"
  | "AEASBD"
  | "BHS";

export const DeviceType: Record<string, DeviceType> = {
  UNKNOWN: "UNKNOWN",
  BAG_TAG_PRINTER: "BAG_TAG_PRINTER",
  BAG_TAG_FEEDER: "BAG_TAG_FEEDER",
  BAG_TAG_DISPENSER: "BAG_TAG_DISPENSER",
  BOARDING_PASS_PRINTER: "BOARDING_PASS_PRINTER",
  BOARDING_PASS_FEEDER: "BOARDING_PASS_FEEDER",
  BOARDING_PASS_DISPENSER: "BOARDING_PASS_DISPENSER",
  BARCODE_READER: "BARCODE_READER",
  PASSPORT_READER: "PASSPORT_READER",
  MSR_READER: "MSR_READER",
  KEY_PAD: "KEY_PAD",
  ANNOUNCEMENT: "ANNOUNCEMENT",
  FEEDER: "FEEDER",
  DISPENSER: "DISPENSER",
  ILLUMINATION: "ILLUMINATION",
  HEADSET: "HEADSET",
  BIOMETRIC: "BIOMETRIC",
  SCALE: "SCALE",
  CAMERA: "CAMERA",
  INSERTION_BELT: "INSERTION_BELT",
  VERIFICATION_BELT: "VERIFICATION_BELT",
  PARKING_BELT: "PARKING_BELT",
  RFID: "RFID",
  AEASBD: "AEASBD",
  BHS: "BHS",
} satisfies Record<string, DeviceType>;